refactor: contracts 컬럼 수정 및 order FK 정리
- contracts parent_contract_id, version 추가
- submitted_at, canceled_at 추가
- quotes/contracts/contract_items updated_at 제거
- quote/contract 관련 PK 타입 int로 통일
- order FK 매핑 수정

+++ backend/src/main/java/kr/remerge/stylehub/domain/negotiation/entity/Negotiation.java
@@ -19,7 +19,7 @@ public class Negotiation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "negotiation_id")
-    private Long negotiationId;
+    private Integer negotiationId;
 
     @Column(name = "negotiation_type", nullable = false, length = 30)
     private String negotiationType;
